feat(web): redesign biomarker detail page with Overview/Research tabs
Replaces the 1205-line biomarker-page-client.tsx monolith (9 stacked uniform-
card sections) with a route-based shape that mirrors the experiment detail
page. New /biomarkers/[id] is Overview (header + About 3-column + image-led
experiment cards with signal-direction hero); /biomarkers/[id]/research
holds claims, sources, and the commons memo.

Lifts SectionLabel to ui/ and trend-comparison helpers to browser-vault/
so both surfaces share the same primitives. BrowserVaultProvider moves into
the new biomarker-layout-client so trend state survives tab switches.
Aliased slug redirects (e.g. spo2 -> blood-oxygen-spo2) now happen on each
leaf page so /spo2/research preserves the Research tab on redirect.

Three temporary frontend lookups carry data fields the model will eventually
own (signal range/window, biomarker about prose, ranking image). Each file
is commented TEMPORARY and follow-up is logged in TODOS.md for @rocketman-21.

Verification:
- pnpm typecheck, pnpm lint clean
- pnpm vitest apps/web/test/: 1214/1214 pass after migrations
- visual readback at /biomarkers/resting-heart-rate (Overview + Research)
- /experiments/finnish-sauna unchanged after SectionLabel lift
- task-finish-review codex pass: 1 medium fixed (alias redirect preserves
  /research), 1 low fixed (exhaustive switch in formatExpectedDirection),
  next-env.d.ts dev artifact restored. Remaining medium (lost model fields
  from old layout) is the design-doc product decision; low test gap deferred.
